添加服务完整版

+++ application/pick/controller/Index.php
@@ -23,22 +23,65 @@ class Index extends Controller
  
 	// * 登录验证 CheckLogin
 	//     - 请求: name/pwd
-	//     - 响应
-	//         + err
-	//         + token
-    public Function CheckLogin($name, $pwd){
-
-    	// header('Access-Control-Allow-Origin:*');
+    public Function CheckLoginTest($name, $pwd){
 
 		$data = ['err' => 0];
 
 		// var_dump(Building::where("id",1)->find());
-
 		$user = User::where('name', $name)->find();
 		if(!$user){
 			$data["err"] = 1;
 			return json($data);
 		}
+
+		if(md5($pwd) != $user->pwd){
+			$data["err"] = 1;
+			return json($data);
+		}
+
+		$data["token"] = "5b18ac6ebbba1";
+
+        UserToken::create([
+            'user_id'  =>  $user->id,
+            'token' =>  $data["token"]
+        ], ['user_id','token'], true);
+
+		$favorite_building_list = Favorite::where("user_id", $user->id)->select();
+		$data["favorite_building_list"] = $favorite_building_list;
+		
+		$order_building_list = Order::where("user_id", $user->id)->select();
+		$data["order_building_list"] = $order_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user->id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
+		}
+
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;		
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
+
+        return json($data);
+    }
+
+	// * 登录验证 CheckLogin
+	//     - 请求: name/pwd
+    public Function CheckLogin($name, $pwd){
+
+		$data = ['err' => 0];
+
+		// var_dump(Building::where("id",1)->find());
+		$user = User::where('name', $name)->find();
+		if(!$user){
+			$data["err"] = 1;
+			return json($data);
+		}
+
 
 		if($pwd != $user->pwd){
 			$data["err"] = 1;
@@ -52,15 +95,14 @@ class Index extends Controller
             'token' =>  $data["token"]
         ], ['user_id','token'], true);
 
-
-		$user_batch_list = UserBatch::where("user_id", $user->id)->select();
-		$data["user_batch_list"] = $user_batch_list;
-
 		$favorite_building_list = Favorite::where("user_id", $user->id)->select();
 		$data["favorite_building_list"] = $favorite_building_list;
 		
 		$order_building_list = Order::where("user_id", $user->id)->select();
 		$data["order_building_list"] = $order_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user->id)->select();
+		$data["user_batch_list"] = $user_batch_list;
 
 		$batch_id_list = [];
 		foreach($user_batch_list as $key=>$user_batch){
@@ -69,6 +111,9 @@ class Index extends Controller
 
 		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
 		$data["batch_list"] = $batch_list;
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
 
         return json($data);
     }
@@ -98,14 +143,14 @@ class Index extends Controller
 			return json($data);
 		}
 
-		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
-		$data["user_batch_list"] = $user_batch_list;
-
 		$favorite_building_list = Favorite::where("user_id", $user_id)->select();
 		$data["favorite_building_list"] = $favorite_building_list;
 		
 		$order_building_list = Order::where("user_id", $user_id)->select();
 		$data["order_building_list"] = $order_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
 
 		$batch_id_list = [];
 		foreach($user_batch_list as $key=>$user_batch){
@@ -115,6 +160,8 @@ class Index extends Controller
 		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
 		$data["batch_list"] = $batch_list;
 
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
         return json($data);
     }
 
@@ -140,13 +187,36 @@ class Index extends Controller
 			return json($data);
 		}
 
+		$favorite = Favorite::where("user_id", $user_id)->where("building_id", $building_id)->find();
+		if($favorite){
+			$data["err"] = 1;
+			return json($data);
+		}		
+
         Favorite::create([
             'user_id'  =>  $user_id,
             'building_id' =>  $building_id
         ], ['user_id','building_id'], true);
 
+		$building->favorite = ['inc', 1];
+		$building->save();
+
 		$favorite_building_list = Favorite::where("user_id", $user_id)->select();
-		$data["favorite_building_list"] = $favorite_building_list;        
+		$data["favorite_building_list"] = $favorite_building_list;   
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
+		}
+
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;	  
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;  
         return json($data);
     }
 
@@ -158,11 +228,39 @@ class Index extends Controller
 			$data["err"] = 1;
 			return json($data);
 		}
-		
+		$building = Building::where("id", $building_id)->find();
+		if(!$building){
+			$data["err"] = 1;
+			return json($data);
+		}
+
+		$favorite = Favorite::where("user_id", $user_id)->where("building_id", $building_id)->find();
+		if(!$favorite){
+			$data["err"] = 1;
+			return json($data);
+		}	
+
 		Favorite::where("user_id", $user_id)->where("building_id", $building_id)->delete();
 
+		$building->favorite = ['dec', 1];
+		$building->save();
+
 		$favorite_building_list = Favorite::where("user_id", $user_id)->select();
-		$data["favorite_building_list"] = $favorite_building_list;        
+		$data["favorite_building_list"] = $favorite_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
+		}
+
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
         return json($data);
     }
 
@@ -185,57 +283,80 @@ class Index extends Controller
 		// 看看房子存在不
 		$building = Building::where("id", $building_id)->find();
 		if(!$building){
-			$data["err"] = 1;
+			$data["err"] = 2;
 			return json($data);
 		}
 
 		// 每期只能一个订单
 		$order = Order::where("user_id", $user_id)->where("batch_id", $building->batch_id)->find();
 		if($order){
-			$data["err"] = 2;
+			$data["err"] = 3;
 			return json($data);
-		}		
+		}	
 
-		// 添加订单;
+		$user_batch = UserBatch::where("user_id", $user_id)->where("batch_id", $building->batch_id)->find();	
+		if(!$user_batch){
+			$data["err"] = 4;
+			return json($data);
+		}	
+
+		$batch = Batch::where("id", $building->batch_id)->find();	
+		if(!$batch){
+			$data["err"] = 5;
+			return json($data);
+		}	
+
+		// 时间判定
+		$begin_time = $batch->begin_time;
+		$end_time = $batch->end_time;
+		if($user_batch->is_preference==1){
+			$begin_time = $batch->preference_time;
+		}
+
+		$begin_time = strtotime($begin_time);
+		$end_time = strtotime($end_time);
+		$now = time();
+		if($now < $begin_time){
+			$data["err"] = 6;
+			return json($data);
+		}
+
+		if($now > $end_time){
+			$data["err"] = 7;
+			return json($data);
+		}
+
+		// 添加订单
         Order::create([
             'user_id'  =>  $user_id,
             'building_id' =>  $building_id,
             'batch_id' =>  $building->batch_id
         ], ['user_id','building_id', 'batch_id'], true);
 
- 		$order_building_list = Order::where("user_id", $user_id)->select();
-		$data["order_building_list"] = $order_building_list;       
-        return json($data);
-    }
+        $building->user_id = $user_id;
+        $building->save();
 
-	// * 房子数据 GetBuilding
-	//     - 请求: token / batch_id
-	//     - 响应
-	//         + err
-	//         + user_data
-	//             * building_list
-	//                 - 房子编号 id，批次编号 batch_id，名字 name，是否车位 is_parking，面积 area，几房几厅几卫描述 desc，单价 unit_price，总价 total_price，订购者编号 book_user_id，收藏计数 favorite_count
-    public Function GetBuilding($token, $batch_id){
-    	
-		$data = ['err' => 0];
-		$user_id = $this->_GetUserID($token);
-		if($user_id == 0) {
-			$data["err"] = 1;
-			return json($data);
+ 		$order_building_list = Order::where("user_id", $user_id)->select();
+		$data["order_building_list"] = $order_building_list;   
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
 		}
 
-		$building_list = Building::where("batch_id", $batch_id)->select();
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;    
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
 		$data["building_list"] = $building_list;
         return json($data);
     }
 
 	// * 收藏数据 GetFavorite
 	//     - 请求: token
-	//     - 响应
-	//         + err
-	//         + user_data
-	//             * favorite_building_list
-	//                 - 房子编号 id，批次编号 batch_id，名字 name，是否车位 is_parking，面积 area，几房几厅几卫描述 desc，单价 unit_price，总价 total_price，订购者编号 book_user_id，收藏计数 favorite_count
     public Function GetFavorite($token){
     	
 		$data = ['err' => 0];
@@ -247,16 +368,25 @@ class Index extends Controller
 
 		$favorite_building_list = Favorite::where("user_id", $user_id)->select();
 		$data["favorite_building_list"] = $favorite_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
+		}
+
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;		
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
         return json($data);
     }
 
-	// * 下单数据 GetBookData
+	// * 下单数据 GetOrder
 	//     - 请求: token
-	//     - 响应
-	//         + err
-	//         + user_data
-	//             * book_building_list
-	//                 - 房子编号 id，批次编号 batch_id，名字 name，是否车位 is_parking，面积 area，几房几厅几卫描述 desc，单价 unit_price，总价 total_price，订购者编号 book_user_id，收藏计数 favorite_count
     public Function GetOrder($token){
     	
 		$data = ['err' => 0];
@@ -268,6 +398,20 @@ class Index extends Controller
 
 		$order_building_list = Order::where("user_id", $user_id)->select();
 		$data["order_building_list"] = $order_building_list;
+
+		$user_batch_list = UserBatch::where("user_id", $user_id)->select();
+		$data["user_batch_list"] = $user_batch_list;
+
+		$batch_id_list = [];
+		foreach($user_batch_list as $key=>$user_batch){
+		    $batch_id_list[] = $user_batch->batch_id;
+		}
+
+		$batch_list = Batch::where("id", "in", $batch_id_list)->select();
+		$data["batch_list"] = $batch_list;		
+
+		$building_list = Building::where("batch_id", "in", $batch_id_list)->select();
+		$data["building_list"] = $building_list;
         return json($data);
     }
 }
